Update smi package name to amdrocm-amdsmi
Refactor:  Extract shared logic into packaging_utils.py
